ai/requirements.txt
rq==1.16.2
redis==5.0.8
boto3==1.35.43
pydantic==2.9.2
pytesseract==0.3.13
Pillow==10.4.0
ocrmypdf==16.9.0
pdfplumber==0.11.4
PyMuPDF==1.24.10
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
